Refactor annots search code into separate class
- AnnotationsStorage class was a bit bloated
- will eventually move this code to be storex plugin
